requirement.txt
aiohttp==3.8.1
aiohttp-retry==2.8.3
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.0.1
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.4
databricks-cli==0.17.3
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docker==5.0.3
docopt==0.6.2
dpath==2.0.6
dulwich==0.20.45
dvc==2.21.2
dvc-data==0.1.23
dvc-http==2.19.0
dvc-objects==0.1.16
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.10.0
entrypoints==0.4
filelock==3.8.0
Flask==2.2.2
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.3.1
fsspec==2022.7.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
greenlet==1.1.3
gunicorn==20.1.0
huggingface-hub==0.9.1
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
mlflow==1.28.0
multidict==6.0.2
nanotime==0.5.2
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pipreqs==0.4.11
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
pycparser==2.21
pydot==1.4.2
pygit2==1.10.0
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
querystring-parser==1.2.4
regex==2022.8.17
requests==2.28.1
rich==12.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.0
scmrepo==0.0.25
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
tabulate==0.8.10
threadpoolctl==3.1.0
tokenizers==0.12.1
tomlkit==0.11.4
torch==1.12.1
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
websocket-client==1.4.0
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.1
